requirements.txt
astroid==2.4.0
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.3
pandas==1.0.3
pylint==2.5.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
toml==0.10.0
wrapt==1.12.1
